Floats and doubles are working.  long doubles?  Not so much.

+++ src/JSTUtils.h
@@ -20,6 +20,6 @@ NSString *JSTStringForFFIType(ffi_type* type);
 ffi_type* JSTFFITypeForBridgeDeclaredType(NSString *type);
 JSValueRef JSTMakeJSValueWithFFITypeAndValue(ffi_type *type, void *value, JSTBridge *bridge);
 long JSTLongFromValue(JSTBridge *bridge, JSValueRef value);
-double JSTDoubleFromValue(JSTBridge *bridge, JSValueRef value);
+long double JSTDoubleFromValue(JSTBridge *bridge, JSValueRef value);
